requirements.txt
git+https://github.com/tandemdude/sqlc-asyncpg-compat
asyncpg~=0.30.0
fastapi[all]~=0.115.4
passlib~=1.7.4
bcrypt==3.2.0
PyJWT~=2.9.0
uvicorn~=0.32.0
python-multipart
python-dotenv~=1.0.1
pydantic~=2.9.2
aiogram~=3.15.0
Pygments~=2.18.0
watchfiles~=0.24.0
Jinja2~=3.1.4